moves around a few components; fixes DatePicker animation issue
