+++ src/components/icons/Keyboard.svelte
