+++ src/components/icons/Slack.svelte
