+++ src/components/icons/ThreeByThree.svelte
