+++ src/components/interactions/Scrub.svelte
