+++ src/stores/options.json
@@ -49,7 +49,7 @@
     "values": [
       {
         "itemType": "section",
-        "label": "Firefox (Desktop)"
+        "label": "Firefox Desktop"
       },
       {
         "label": "Any Firefox Desktop Activity",
